Created an abstract Application class with application Layers in a LayerStack
The abstract application class is meant to be used as a basic application to open a glfwWindow and update the application layers.

The layers can be used to seperate the game logic from the (later) ImGui logic.
By differentiating between a layer and an overlay in the layerstack, we have control over what gets rendered/updated first. This is gonna be important when we implement all the ImGui context later on, because ImGui has to be rendered on top of everything.